RCU: Capture performance results on Linux and QNX
Provide a toolchain and enable compilation on QNX (tested with QNX 7.1.0 and QNX 8.0.0).

Move the stress test code from being a test case to the binary `main.cpp` which is compiled as `src/rcutest`.

Results ran on a Raspberry Pi4B (ARM A72, 4 core) with QNX 7.1 and 8.0 and Linux. Raspberry Pi 5 (ARM A76) on Linux.

Results ran on an older Intel i7-6700T and a newer Xeon Silver 4410Y.

Issue: OSDEV-10