Cache Line: Measure rates for copying strided
Reimplement using Google-Benchmark the same results from Daniel Lemire [1], and the results parsed using a Python Script and GnuPlot to generate the same results.

The size of the buffer is configurable (defaults to 256MB) as if it is too small, the L3 cache is measured instead of the memory bandwidth. Use the `-b` option to configure the buffer size (see the examples in the results directory).

In the file `cacheline.md` we describe how it works.

Provide results of the benchmark:
- i9-13980HX (using 256MB array size, as it has a L3 cache of 36MB).
- i3-2120T (using 8MB array size, L3 cache of 3MB).
- i7-4930K (using 32MB, L3 cache of 12MB).
- RPi4 (using 8MB, L3 cache of 1MB)
- RPi5 (using 8MB, L3 cache of 2MB)
- Intel Core2 Duo T-7700

[1] https://lemire.me/blog/2023/12/12/measuring-the-size-of-the-cache-line-empirically/

Issue: OSDEV-24