Core Latency: Test and optimise for Raspberry Pi4/5
Investigated the assembly generated by the compiler and was able to optimise it by 10% when running on Rasbperry Pi 5 by hand-writing the assembly. By having hand-written assembly, we ensure that the comparisons are fair between architectures, instead of it testing the feature set and possibly using the `cas` instruction instead.

Future work could include testing for the FEAT_LSE of ARM and implementing the `cas` instruction.

Issue: OSDEV-23