Google Benchmark: Disable tests always
Google-Benchmark compiles against C++14, and it expects when it uses the provided version of Google-Test, that it is also compatible with C++14. Unfortunately, the version actually used in this project is newer than what Google-Benchmark wants, and will fail compilation in some cases.

In particular, when compiling with Clang 10.0 or later, it will complain about the usage of:

```
benchmark_gtest.cc:15:1: error: use of the 'maybe_unused' attribute is a C++17 extension [-Werror,-Wc++17-attribute-extensions]
```

That is because the version of Google-Test really does use this, but it isn't allowed with C++14.

Two options:

1. Downgrade the version of Google-Test to be compatible (Google-Benchmark builds locally with commit `e2239ee6`)
2. Disable testing.

I don't want to downgrade to C++14. So disabling tests appears to be the best option.

The "correct" solution would be to somehow force Google-Benchmark to actually compile with C++17. However the variable `BENCHMARK_CXX_STANDARD` is not a cached variable, and it is set to 14. By compiling Google-Benchmark as its own library with C++14, we hope that the compiler vendor doesn't change the ABI when linking into a program (like `cacheline_bench`) that is compiled against C++17.

QNX fails to build if `set(CMAKE_CXX_EXTENSIONS OFF)` is set. So enable it by default, but not for QNX.

Issue: OSDEV-30