Google Benchmark: Set the libraries to compile with C++17
Set the libraries that we link to, `benchmark` and `benchmark_main`, to use features of C++17 that it is ABI aligned with the rest of the project. Confirmed with `VERBOSE=1 make`, that the libraries really do compile with the option `-std=c++17` on GCC and Clang on Ubuntu 22.04.

Note, the test code will still compile against C++14, and so will still not build using Clang, because the default is C++14. We're only overriding for two libraries which our other code links to. Hence, `set(BENCHMARK_ENABLE_TESTING OFF)` must remain as is.

Issue: OSDEV-30