String Comparison: Benchmark C++17 and LibC routines
Compare the various ways where one can check if a string starts with a particular substring. In C++17 there are three different ways, and there's the `std::string` and `std::string_view`.

We observe that in most cases, if the substring is longer than the string itself, it makes sense to do a length comparison, even if the results are the same.

Results are posted for Linux (Intel i9-13950HX, RPi4, RP5) and QNX 7.1 and 8.0 (RPi4).

All tests were done with GCC.

Issue: OSDEV-32