Malloc Bench: Rewrite tests for Linux and QNX
Test cases for `malloc_bench` were completely rewritten to provide the ability to test different features on how an operating system typically allocates memory through the API `mallopt()` and `mlockall()`. Support for `mallopt()` is baked in for:

- Linux
- QNX 7.1.0 (as per https://www.qnx.com/developers/docs/7.1/#com.qnx.doc.neutrino.lib_ref/topic/m/mallopt.html)
- QNX 8.0.0 (as per https://www.qnx.com/developers/docs/8.0/com.qnx.doc.neutrino.lib_ref/topic/m/mallopt.html?hl=mallopt)

See `CMakeLists.txt` on how it detects the options if you want to add more (it's not automated).

Tests are executed on the following platforms:

- Intel 13th generation i9-13950HX with Ubuntu 22.04, Kernel 6.8.0
- Raspberry Pi 4, RPi OS 5.3 (Debian)
- Raspberry Pi 4, QNX 7.1.0
- Raspberry Pi 4, QNX 8.0.0
- Raspberry Pi 5, RPi OS 5.3 (Debian)

The tests run look for:

- speed of `malloc()` and `free()`
- speed of `calloc()`
- any variations of performance when walking the pages (looking for memory that might result in a page fault)
- speed of memset(0), and walking, looking for zero-page behaviour

A Jupyter notebook (analyse.ipynb) is created that makes it easier to read the Google Benchmark `.json` file outputs and parse them as histograms side-by-side between different test runs (with different parameters). This makes it easier to identify the performance differences between the options. A user can visually see how their code might be impacted.

QNX tests were run without superlock and with superlock (the procnto-smp-instr `-mL` option). Then it is easier to compare against three flavours of Operating System all running on the same hardware.

To use the tool, run:

```sh
$ malloc_bench -?
USAGE: ./benchmarks/malloc/malloc_bench [-mOPTION=n] [-L]

Run a malloc benchmark.

 -L: Enable locking with mlockall().
 -mOPTION=value: a comma separated list of mallopt options and a value.

Options supported by '-m' are:
 - M_ARENA_MAX
 - M_ARENA_TEST
 - M_MMAP_MAX
 - M_MMAP_THRESHOLD
 - M_MXFAST
 - M_PERTURB
 - M_TOP_PAD
 - M_TRIM_THRESHOLD

benchmark [--benchmark_list_tests={true|false}]
          [--benchmark_filter=<regex>]
          [--benchmark_min_time=`<integer>x` OR `<float>s` ]
          [--benchmark_min_warmup_time=<min_warmup_time>]
          [--benchmark_repetitions=<num_repetitions>]
          [--benchmark_enable_random_interleaving={true|false}]
          [--benchmark_report_aggregates_only={true|false}]
          [--benchmark_display_aggregates_only={true|false}]
          [--benchmark_format=<console|json|csv>]
          [--benchmark_out=<filename>]
          [--benchmark_out_format=<json|console|csv>]
          [--benchmark_color={auto|true|false}]
          [--benchmark_counters_tabular={true|false}]
          [--benchmark_context=<key>=<value>,...]
          [--benchmark_time_unit={ns|us|ms|s}]
          [--v=<verbosity>]
```

The tests try to make sure as little use of the heap as possible, by overriding the new/delete operators to a static amount of memory (8MB). The amount of memory used is printed at the end of the program. If for some reason, the test faults too early, it might be that the memory region allocated is not enough (it never frees to make it as simple as possible).

It has been compiled for the MacOS-X (M3 chipset), but the results all return zero, indicating that the test doesn't run properly. So don't use on a MAC.

Issue: OSDEV-28