Core Latency: ARM tests with LLSC and LSE
Provide the ability to test different implementations of the ping/pong algorithm on processors.

On all processors, the default uses the C++ implementation and what code the compiler generates. This can be compared against hand-crafted assembly code in other tests.

On Intel (32-bit and 64-bit), the `-bcas` runs the hand-written assembly.

On ARM, the `-bllsc` runs the hand-written assembly using the exclusive load / store instructions that require a memory monitor. This works on all Aarch64 targets.

For newer ARM64 processors, one can enable the LSE instruction set, to test with the `cas` assembly instruction. To do this, the compiler options must be given that the assembler can generate the code and understand the `cas` instruction. For example, give the option `-DCMAKE_CXX_FLAGS="-march=armv8.1-a+lse"` to get this feature. Then the test can be executed with `-blse`. A test is made on the board for Linux and QNX that the processor actually supports these instructions. Note, by enabling the optimisations, the compiler may automatically generate code using illegal instructions on hardware if it doesn't support LSE.

The second major update of this change is to implement a "statistics" module for more reliable measurements. The `-s` option is the number of samples, which each sample contains a set number of iterations given by `-i`. The output is now the "median" of the measurement, instead of the average, providing a closer representation of the true result.

Issue: OSDEV-36