CMake QNX: Remove C dependency on QNX check
Use the `ENABLED_LANGUAGES` property to know if the test should be run using a C or C++ compiler. Then we test for C first, then followed by C++ if needed.

This now allows a project to configure only for CXX if needed (and drop the C project dependency, which was needed because of this test).

It is up to the toolchain maintainer to ensure that the two compilers are correctly configured.

Issue: OSDEV-42