libubench: Move Threadpin here
We're likely going to start needing common code. Take the first Operating System dependent code and move it to this library.

The signature is updated to return success instead of aborting, so that code using the library can decide what to do.

On MacOS-X, there is no way to pin a thread. So provide an empty implementation.

Issue: OSDEV-43