libubench: Implement an Idle Clock for Linux/QNX
Provide idle_clock, which provides the absolute amount of idle time in the system across all cores at this point in time.

Under Linux, we must use `/proc/stat` to read the total idle time, and then convert it to a `time_point` based on the number of jiffies per second (the clock tick). If we ever find a mechanism to get a higher resolution clock, the code can be extended to implement a new class and provide a better clock if available.

Under QNX, read the clock times for the idle threads from procnto.

Issue: OSDEV-43