udp_load: Send UDP traffic stress test
Provide a tool that sends traffic data out via UDP packets from a source interface, to a destination address.

The shaping is tested via a simulation in the `docs` folder, described in a Jupyter notebook that shows how the shaper can work. This was done before being implemented in C++ for quick observation of how the shaper might work and to detect possible errors depending on timing.

Issue: OSDEV-43