Ubuntu Podman Images
Provide a Dockerfile for creating a Podman image for different Ubuntu releases based on the CODENAME.

The `build.sh` script can be run that:

- Builds a podman container image from an Ubuntu CODENAME and installs the necessary tools inside.
- Runs the shell in the container in interactive mode
- Allows running as root (not the default)
- Allows running a specific command

For example:

 ```cmd
 jcurl@localhost$ build.sh -cjammy -i
 jcurl@180ae061e343$ cmake -B . -S /source/qnx -DCMAKE_BUILD_TYPE=Release
 jcurl@180ae061e343$ make -j8
 jcurl@180ae061e343$ make clangformat
 ```

Or to build on a single command line

 ```cmd
 $ build.sh -cjammy "cmake -B . -S /source/qnx -DCMAKE_BUILD_TYPE=Release && make -j8"
 ```

Output is put in the folder `qnx/build/ubuntu-CODENAME`.

The second way is to use the `Makefile`

 ```cmd
 $ cd qnx/scripts
 $ make
 ```

This will build the containers and build the software inside the containers for GCC and Clang. This helps automate checking the software before a commit.

Preconditions:

- Ensure that you've set up the `/etc/subuid` and `/etc/subgid` file first to allow podman to run as non-root.

Building the container removes unnecessary files to keep it small as possible. If you want to test providing updates, you will need to run `apt update` first, to populate the repository. Note, that running in `root` mode will let you modify the container, but the `build.sh` script uses the `--rm` option to delete the container when finished, so your results are not persisted.

Issue: OSDEV-49