docker: Add alpine-docker Dockerfile
Add a pure LLVM toolchain in addition for the Ubuntu targets.

By compiling for multiple targets, the latest technologies can be used to improve code quality. By using a pure LLVM toolchain, further warnings and analytics are provided allowing for improved build quality.

The `build.sh` script is updated to support a distribution, as well as looking for distribution specific files. With LLVM it was necessary to have distribution specific docker files as the names of the packages required for installation differ slightly between distributions.