CMake Refactoring
Move custom CMake modules into the folder `research`, to separate new modules defined by this project with possible external implementations that might need updating.

Use `CMAKE_CXX_COMPILER_LOADED`, which is also done internally by KitWare CMake macros, instead of checking the language list explicitly. This shortens the code.

The QNX macros were updated, to additionally set the `CMAKE_SYSTEM_VERSION` in the `PARENT_SCOPE`. Note, while it is also set as `CACHE INTERNAL`, this is not enough to allow Google Test and Google Benchmark to work. So make sure that the `research/qnx` is included at the top level of your `CMakeLists.txt` file. This has allowed to remove the checks for QNX in the subprojects because GoogleTest automatically adds `-lregex` if it sees version 7.1.0 or later. Dependencies to source files were removed, as the function now creates it dynamically, making it more portable for reuse.

The `arm64_has_lse` is updated to cache properly results to not rerun if executed multiple times (especially if ARM is not the target toolchain).

The `check_mallopt_option_exists` now only runs a check once, if run multiple times. It will only compile for C if the project is loaded, or then only for C++.

The `find_libsocket` was reworked to support caching and set the variables from previously cached data.

Issue: OSDEV-51