NetBSD: Allow basic compilation for cross compile NetBSD 10.1 to RPi4
Provide a sample toolchain file for CMake, based on a cross-compile toolchain for Linux (locally built). The `README.md` contains some very brief instructions on building a toolchain (which turns out to be quite simple).

NetBSD 10.x supports `pthread_setaffinity_np` but it is not compatible with Linux. Add an explicit implementation as per the manpage.

Support is missing for measuring CPU times and getting the process names and must be updated in subsequent patches.