libubench: Add test cases checking for support
When porting to a new Operating System, it is useful to know what must be ported. For example, porting to NetBSD requires new implementations in clock, threads and process names. Rather than test each individual test case, the implementation itself can tell us if it is a "null" implementation (not supported) or not.

The test case `ubench_test` now fails on functionality indicating it must be ported.

Issue: OSDEV-59