podman: Update alpine-latest-clang
Building new a new Alpine fails on the docker image generation with the error:

 ```
 ERROR: unable to select packages:
   libunwind-dev-1.8.1-r0:
     breaks: llvm-libunwind-19.1.4-r0[!libunwind-dev]
     satisfies: world[libunwind-dev]
 ```

Use the package `llvm-libunwind-dev` instead.

The `build.sh` is updated to allow read/write access to the `/source` folder. An additional check is made if building with no commands in non-interactive mode (it can still be used to build the container).