FreeBSD: Implement idle clock
Implemented capturing the IDLE CPU time for the entire system on FreeBSD. On FreeBSD there are no constants to query `sysctl`, so they need to be mapped first. Otherwise the implementation is the same as for NetBSD.

The `scripts/build.sh` now enables the source directory to be read/write when in interactive mode. This allows running `make clangformat` within the container. The automatic building still remains read-only.

Refactored that `base_clock` is a common implementation.

The tool `udp_load` was tested that uses this implementation and seems to provide accurate results.

Issue: OSDEV-63