libubench: Utility functions for Ethernet network addresses
Provide the following utility classes

- `is_multicast` - test IPv4 if it is a multicast address.
- `get_ether_multicast` - Convert an IPv4 multicast to Ethernet address.
- `inet_ntos` - Convert a `sockaddr_in` to a `std::string`
- `ether_ntos` - Convert a `ether_addr` to a `std::string`.
- Stream operators to print IPv4 and Ethernet addresses.
- `query_net_interfaces` - Get a list of all interfaces in the system.
- `query_net_interface` - Get details of a specific interface.

The classes `if_ipv4` and `interfaces` which is used by `query_net_interfaces` is chosen that the fields are read/write. Initially, it was implemented so that the class is immutable and needed friends for construction. The better solution however, is to allow a mutable class through setters and getters, and the method that returns the list of interfaces is returned as `const` so that it can no longer be modified. Benefits include reducing the need for friends (implementation details), and simplifying the implementation of the code constructing the objects necessary.

A `std::vector` is used to maintain the list of addresses assigned per interface. This implementation may need to change to support IPv6 and other address types per interface. Important is that the interface addresses can be enumerated in an easy manner, with the possibility to easily filter for each address type.

When using `getifaddrs`, the internal implementation uses a simple implementation of an `input` iterator, making the management of the object easy and freed by the destructor.

On Linux, Cygwin, we get the hardware address, if not all zero, with `SIOCGIFHWADDR`. On Linux, the struct ::ether_addr exists from `net/ethernet.h`. On Cygwin, we don't have this, and have to use `sockaddr` directly.

On NetBSD, QNX 6.4.1 and later [1], the `getifaddrs()` can return the MAC address information as part of `AF_LINK` family type. This requires the struct `sockaddr_dl` and `IFT_ETHER`. This code only supports Ethernet MAC addresses.

To check for types, the CMake macro `check_type_exists()` is implemented that is similar to how `check_symbol_exists()` and code from CMake 3.16.9 was used, along with the logic of type checking from Autoconf [2]. This is used to check for the existence of `sockaddr_dl`.

On NetBSD and QNX, values like IFT_ETHER are a `#define`. But on QNX 8, it's a value in an enum. Implemented `check_symbol_enum_exists()` to first check for the symbol, and if it fails, then check if it is an integer type.

VLAN identifier and the parent interface are obtained. For this to work on Linux, one must have the linux headers. Cygwin doesn't expose functionality to get VLAN information about the interfaces.

The `ioctl()` implementations use the BSD `strlcpy` to ensure strings are properly nul-terminated (which `strncpy` doesn't guarantee). For BSD systems, this is correct. For Linux, interfaces that have a length of IFNAMSIZ (without a nul-terminator) therefore won't be detected.

Issue: OSDEV-47

[1] https://www.qnx.com/support/knowledgebase.html?id=50130000000mosy
[2] http://git.savannah.gnu.org/gitweb/?p=autoconf.git;a=blob;f=lib/autoconf/types.m4;h=508b64a21e159725e0233b7289053eb77ec1dc43;hb=HEAD