libubench: Create strings namespace and expose strlcpy
Other code will need to do OS specific functionality directly, especially `ioctl` where it makes sense to use `strlcpy`. This was available internally for use by libubench, the functionality should also be exposed. Use the opportunity to rename the `args` namespace to be closer to what it really is - string manipulation.

Issue: OSDEV-68