udp_load: Run measurements
Rerun on the following targets and capture the outputs:

- QNX 7.1 (io-pkt) for sendto, sendmmsg, bpf
- QNX 8.0.2 (io-sock) for sendto, sendmmsg, bpf
- FreeBSD 14.2 for sendto, sendmmsg, bpf
- NetBSD 10.1 for sendto, sendmmsg (bpf broken)
- Linux RPiOS RPi4 for sendto, sendmmsg
- Linux RPiOS RPi5 for sendto, sendmmsg

On BSD Operating Systems the methods `sendto`, `sendmmsg` and `writev` may return ENOBUFS, where Linux doesn't do this. We now handle this case, by delaying a short period of time. On FreeBSD we can use `CLOCK_REALTIME_PRECISE`. Unfortunately, there's no mechanism that I could find on these OSes to know what the optimal amount of time to wait for is. The number chosen (750us) is arbitrary just to get it to work. QNX 7.1 and QNX 8.0 fortunately do not show this behaviour either.

Issue: OSDEV-69