udp_talker: fix leak of FD on error in bpf
The ubench::file::fdesc class move assignment operator no longer closes the FD when it moves a file descriptor it already has (assigning to itself).

When moving to use RAII with ubench::file::fdesc, determined that on init of the BPF test, if there was an error, the file descriptor remained open.

Now the udp_talker implementation is slightly more robust due to the RAII pattern.