libubench test: Ensure to get a non-zero value for process_clock::now
On NetBSD/i386 10.0, the clock resolution for the process clock is so low, that while code is executing, it might not be enough to cause the tick counter to increment. This results in the clock returning zero CPU time executed, thus the test case failing.

Burn some CPU, with an upper limit, to see that the clock will tick over, and thus be non-zero.

Issue: OSDEV-76