libosqnx: Move PIDs to a common library
The upcoming `lshmem` will need to iterate over the PIDs in the system, which is already implemented in `lsqf`. Take the implementation and make it common for QNX. While we could make it common for all (and put it in libubench), there's no need for that (yet). When we need it for Linux/BSDs, we'll do that work.

Issue: OSDEV-54