clang-format
The target `clangformat` is only created if the `clang-format` binary is actually found. If you uninstall the binary without remaking the entire project, then the targets will remain.

The `CMakeLists.txt` were updated for the new name of the function.

The `CLANGFORMAT_FOUND` is now cached, that checks are not repeated between runs. The user can provide `CLANGFORMAT_EXECUTABLE` to override the default of `clang-format`.

Can specify the minimum version of clang-format. For this project, it is 10.0 (the version available in current Cygwin is 8.0.1 and is too old, resulting in warnings).

One can still only provide the function `target_clangformat()` once, but can give various options:

- TARGET <target>: The name of the existing target, or a new target
- DIR <dir>: A list of directories to glob and format
- DIR_RECURSE <dir>: A list of directories to recurse to glob and format
- SOURCE <source>: A list of sources to format

With this implementation, some files were not properly formatted (not chosen by the build system for whatever reason) and are now correclty formatting. Only files in folders (usually targets) that are chosen are still not formatted. As an example, QNX targets won't be formatted on Linux, because they're not even built.

Issue: OSDEV-26