str_intern: Benchmark very simple solution
Prepare the environment for being able to test and benchmark a possible `str_intern` implementation.

This implementation is deliberately poor, because it's very fast to implement and therefore test.

Move the 'udp_load' implementation of `busy_measurement` to `ubench::measure` as it is useful for this tool also.

Issue: OSDEV-91