str_intern: var_set peak memory reduction and load factor
Similar to a fixed set, but if the load factor is 1.0 or higher, then double the number of buckets and rehash.

When rehashing, reset the linked list in each hash bucket while rehashing, to help reduce the total amount of memory in the measurements.

Provided a max_load_factor. The default is 1.0, but an implementation might want to use a load factor of greater than 1.0 to reduce the memory footprint at the cost of speed when searching. With a max_load_factor of 1.0, there is no noticable change in performance. To know the impact of the load factor, compare the table parameters with the fixed_set_XXX to get an estimate in the increase in time.

Issue: OSDEV-92