str_intern: Port to ubench library
The implementation of `var_set_pmr` was copied to `libubench`.

Add unit tests. Testing of `max_load_factor` show fixes needed. Original implementation could go out of bounds, but no longer.

The move constructor needed to be fixed, so that the PMR implementation can be moved without invoking undefined behaviour. The PIMPL implementation required the move constructors to not be defaulted in the header file, but in the unit file.

Issue: OSDEV-92