bpf_udp4: PoC for generating UDPv4 packets
The PoC shows how one might want to send UDPv4 packets over the wire using the BPF interface. This is an example where we construct the entire L2 frame from scratch, without any support from a IPv4 stack. UDP packets up to the size of 16-bit are supported, sending out from the interface you specify. Because we're constructing the packet by hand, no ARP is available, and only multicast destination addresses are supported.

Corrected in `udp_load` the checksum calculation for off-sized packets to not access the payload out of bounds.