string: Remove undefined behaviour on from_chars_hex
If the input type is a signed integral type, and the number of characters on the input exceed the size of the type, we'll shift them out expecting the upper bits to be simply removed. In the case of a signed type, this shift invokes undefined behaviour.

So do all the operations on the unsigned version of the integral type and static cast it back to the proper type when done.

Issue: OSDEV-101