rcu: suppress clang-analyzer warning on memory
When building with clang 20.1.1 (on Ubuntu Plucky), there would be false positives that memory is not being freed. This is to be expected, because the `rcu` class and `rcu_ptr` take over the ownership of the memory and manage it explicitly.

Test cases show clearly that memory is being freed (ironically, the test case that is being complained about, on line 380, is where one of the warnings are, and we can see when running the test code that the memory really is being freed).

The `EXPECT_EQ` has no effect (removing those lines does make it easier to analyse though).

For now, as I don't find documentation to tell clang-analyzer the intentions of memory allocation and returns, we simply call `std::abort` when analyzing, that the analyser thinks that the program exits and so no memory checks need to be made.

Issue: OSDEV-104