ubench: Parse short options
Provide a class that has equivalent functionality to `getopt()`.

Many places in code need to suppress lint warnings for the usage of C-like arrays, even if C++ doesn't provide any alternative. This implementation allows to construct a class for parsing the command line options (intended to be very simple) that behaves like existing `getopt()`.

Components may want more control over parsing arguments. For example, new command line option styles have a base tool, then that supports a sub-command, which needs to parse its own options. By providing a vector of arguments, it's easy to manipulate the argument vector (just remove the argument not needed) and parse those arguments. Users of `getopt()` could provide tricks such as updating `optind`, but this makes it potentially easier to read and is not likely to be portable across Posix Operating Systems (see the path `lib/libubench/test/getopt_compat/README.md` for details).

It is intended easy to use, just pass the arguments from your function to the `options` class, and iterate over it. Once you've finished iterating over the options (and handling errors), then just iterate over the args. The arguments are parsed when given to the constructor.

Issue: OSDEV-115