QNX: Add podman images and instructions
Added two build options: `qnx-710-aarch64` and `qnx-800-aarch64`.

The `build.sh` is updated to detect the "distro" of "qnx" and change the options to podman to handle the building.
- This version copies the QNX images into the container. Because of security topics of podman, the installation path of QNX is tarred and then copied into the new container.
- To make it easier to debug with the `-D` option, the tarball is not copied first, but mounted via podman and then extracted. This means when there's an update in the installation path, there is no detection causing the tarball to be updated, and so no automatic update of the container when testing. It was found using COPY or ADD instead that this operation was always slow, regardless of there being no changes to the installation of QNX.

When mounting QNX, the `build.sh` mounts the license directory folder into `/home/user/.qnx` or `/root/.qnx` when running root mode. The docker file now sets the `HOME` directory to be `/home/user` (instead of `/build` for the other containers). It is `/root` when running root mode.

So that the environment variables can be set, there is now a `/home/user/.profile`, which runs `qnxsdp-env.sh` to set up the environment. This requires that running the shell must be given the `-l` option to treat this as a login.

All containers now get a hostname based on the name of the container (based on `-d` and `-c` options, but not `-v` option) and the `/etc/machine-id` file. This allows for consistent builds, so that QNX images provide the HOST in the `use -i` output.

All containers now get the `USER` environment variable set from the host. When running in root mode, the user starting the script is still present, and an indication in the `USER` environment variable the container is running in root mode.

Issue: OSDEV-123