QNX: Podman should use QNX SDP in place
In the previous commit 922534e ("QNX: Add podman images and instructions"), the SDP is copied into the image. It can take up to an hour, and the first run is also slow.

Solve this problem completely by not copying the SDP into the container and mount it instead.

Issue: OSDEV-124