dev-adder: Benchmark devctl calls
This basic driver provides a single `DCMD_ADDER_ADD` command that just takes a `uint32_t` and adds it, driven by a single threaded resource manager. It is to compare the performance of receiving a `_IO_DEVCTL` message, process it and return the result to a single threaded client.

The tool `adder` exercises the interface from `dev-adder` to make performance measurements.

Issue: OSDEV-129