libcpuid: cpuidreader_dev uses now cpuid_basic_ctx
Use the generic implementation, that allows us to remove code also from the header, making the context truly private. The implementation changes slightly, so that when the context is closed, the file handle remains open, until a different context is needed, or until the cpuidreader_dev is destroyed.

Issue: OSDEV-166